refactor model in to correct module
